Trello-KNGQXdJZ: Refactored the way how we do unit tests
- created helper folder where all helper modules will be
- refactored unit tests to remove private/public folders
